refactor(utils): separate utils in different files
