+++ roquiz/lib/view/view_quiz.dart
@@ -7,7 +7,9 @@ import 'package:provider/provider.dart';
 import 'package:roquiz/model/persistence/settings.dart';
 import 'package:roquiz/model/quiz/question.dart';
 import 'package:roquiz/model/quiz/quiz.dart';
-import 'package:roquiz/model/utils/utils.dart';
+import 'package:roquiz/model/utils/device.dart';
+import 'package:roquiz/model/utils/grade.dart';
+import 'package:roquiz/model/utils/time.dart';
 import 'package:roquiz/widget/constrained_appbar.dart';
 import 'package:roquiz/widget/custom_back_button.dart';
 import 'package:roquiz/widget/grade.dart';
